requirements.txt
APScheduler==3.6.3
attrs==19.3.0
cassandra-driver==3.25.0
certifi==2019.11.28
Click==7.0
colorama==0.4.4
colorhash==1.0.2
configparser==4.0.2
contextvars==2.4
cycler==0.10.0
Cython==0.29.24
Flask==1.1.1
Flask-Cors==3.0.8
Flask-MonitoringDashboard==3.0.6
geomet==0.2.1.post1
googletrans==3.1.0a0
gunicorn==20.1.0
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2021.9.1
httpcore==0.9.1
httpx==0.13.3
hyperframe==5.2.0
idna==2.10
importlib-metadata==1.4.0
itsdangerous==1.1.0
Jinja2==2.11.0
joblib==0.14.1
jsonschema==3.2.0
MarkupSafe==1.1.1
matplotlib==3.1.2
more-itertools==8.1.0
nltk==3.6.2
numpy==1.19.5
pandas==0.25.3
Pillow==8.3.2
psutil==5.6.7
pyparsing==2.4.6
pyrsistent==0.15.7
python-dateutil==2.8.1
pytz==2019.3
PyYAML==5.3
regex==2021.8.28
regexp==0.1
rfc3986==1.5.0
scikit-learn==0.22.1
scipy==1.5.3
six==1.14.0
sklearn==0.0
sniffio==1.2.0
tqdm==4.62.3
typing-extensions==3.10.0.2
tzlocal==2.0.0
unicodedata2==13.0.0.post2
Werkzeug==0.16.1
wincertstore==0.2
wordcloud==1.8.1
zipp==2.0.1
gensim==4.1.2
gevent==21.8.0